crawler/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 10.4.0
Markdown == 3.7
PySocks == 1.7.1
PyYAML == 6.0.2
aiohttp == 3.8.4
aiohttp_socks == 0.8.0
beautifulsoup4 == 4.12.3
colorama == 0.4.6
dask == 2023.3.2
diskcache == 5.4.0
docopt == 0.6.2
durations_nlp == 1.0.1
filelock == 3.16.1
html2text == 2024.2.26
html5lib == 1.1
ipython == 8.4.0
joblib == 1.4.2
json5 == 0.9.25
langdetect == 1.0.9
lxml == 5.3.0
matplotlib == 3.7.5
mongoengine == 0.29.1
numpy == 1.24.4
p_tqdm == 1.4.2
pandas == 2.0.3
pathvalidate == 2.5.2
pillow == 10.4.0
playwright == 1.47.0
pymongo == 4.10.0
redis == 5.1.0
requests == 2.32.3
rq == 1.16.2
scikit_learn == 1.1.1
seaborn == 0.13.2
setuptools == 75.1.0
spacy == 3.5.1
titlecase == 2.4
tld == 0.13
tldextract == 3.4.0
tqdm == 4.66.5
user_agents == 2.2.0
xgboost == 1.7.5
